refactor(settings): extract generalPage; fix duplicate/access bugs from prior commit
Completes the TranscriptedSettingsView page extraction (all 10 pages now
live in sibling TranscriptedSettingsView+*.swift files). Also fixes two
real bugs the previous commit shipped, caught by CI (build-and-test failure
on d52b231):

- modelsPage and privacyPage were left declared in both the main file and
  their new extension files (duplicate declaration) after an in-progress
  extraction pass landed mid-commit.
- 5 identifiers Shortcuts.swift needs (sortedAutoEnterAllowedBundleIDs,
  setAutoEnterApp, refreshAutoEnterAppCandidates, chooseAutoEnterApp,
  autoEnterDisplayName) were still `private` in the main file, inaccessible
  from the extension file.

Verified by hand (no local Swift toolchain in this session): brace balance
and single-declaration checks across all 11 files pass. Relying on Swift CI
for the actual compiler check.
